optz: depth only needs a u8
Also adds a TODO to rewrite Move

+++ src/board/movegen.rs
@@ -16,7 +16,7 @@ use std::sync::atomic::AtomicBool;
 impl BoardState {
     pub fn find_best_move(
         &mut self,
-        depth: i32,
+        depth: u8,
         cancellation_token: &AtomicBool,
         debug_mode: &mut bool,
     ) -> Move {
