+++ src/search/lmr.rs
@@ -1,5 +1,5 @@
 #[inline(always)]
-pub fn needs_reduction(depth: i32, number_of_legal_moves: usize) -> bool {
+pub fn needs_reduction(depth: u8, number_of_legal_moves: usize) -> bool {
     if depth < 3 || number_of_legal_moves < 3 {
         return false;
     }
@@ -9,6 +9,6 @@ pub fn needs_reduction(depth: i32, number_of_legal_moves: usize) -> bool {
 
 // TODO: dynamic reduction
 #[inline(always)]
-pub fn get_reduction(_depth: i32, _number_of_legal_moves: usize) -> i32 {
+pub fn get_reduction(_depth: u8, _number_of_legal_moves: usize) -> u8 {
     1
 }
